Move generated rekordbox file to a kaitaistructs folder and disble checks for these generated files
