+++ src/library/rekordbox/rekordboxfeature.cpp
@@ -12,8 +12,8 @@
 #include "library/dao/trackschema.h"
 #include "library/library.h"
 #include "library/queryutil.h"
-#include "library/rekordbox/rekordbox_anlz.h"
-#include "library/rekordbox/rekordbox_pdb.h"
+#include "library/rekordbox/kaitaistructs/rekordbox_anlz.h"
+#include "library/rekordbox/kaitaistructs/rekordbox_pdb.h"
 #include "library/rekordbox/rekordboxconstants.h"
 #include "library/trackcollection.h"
 #include "library/trackcollectionmanager.h"
